Added a minimal sixel example from Chris F. Chiesa
More sixel examples are needed.

This one is dated 1990-09-29, so over 33 years ago and very well documented in https://www.digiater.nl/openvms/decus/vax90b1/krypton-nasa/all-about-sixels.text 

First we see that the introductory  and termination  sequences (lines 1
and  6)  are  in  seven-bit  form,  and  that  no parameters are explicitly
specified in the introductory sequence.   (The absence  of parameters means
that your  device, or  software, will  use its "default" behavior, whatever
that is.)

    Line 2 sets color-map registers 0,  1,  and  2,  all  in  RGB notation.
Color 0  is set  to the  RGB triplet  (0,0,0) --  BLACK.  Color 1 is set to
(100,100,0) -- YELLOW.  Color 2 is set to (0,100,0) -- GREEN.   Color  3 is
unspecified, so will retain whatever setting it had previously.

    Line 3  supplies, first,  the command  to "use color (map register) 1,"
then 14 sixels' worth of image data, then the "$" character which  causes a
"carriage-return" to  the beginning  of this  SAME line -- meaning that the
next data will OVERSTRIKE this line's data.

    Line 4 performs much as did line 3, except that  it specifies  color 2,
and that  it ends  with the  "-" character, meaning that line 5's data will
appear on a new line, rather than being overstruck again.

    Line 5 specifies 14 sixels of value 1 ("@" =  ASCII 64;  64 -  63 = 1),
by using  a repeat-count  rather than  14 individual  instances of the same
character.

    If you haven't figured it out yet, this image produces  a 14-by-7-pixel
yellow rectangle  with the  word "HI"  picked out  on it  in green!  On the
VT240, with its pixel aspect ratio of 1:2, this rectangle  should appear as
a SQUARE, if I haven't screwed up somewhere.  Like this:

    ..............
    ..**..**..**..
    ..**..**..**..        . = color 1, yellow
    ..******..**..        * = color 2, green
    ..**..**..**..
    ..**..**..**..
    ..............

    This should  appear on  a black background.  I'm afraid it'll be almost
too small to see on a VT240 or 340, but you should be at least able to tell
that it's there...

    That about  does it  for the theory end of it; I'll tack on some source
code for your enlightenment and edification, and see  if it  makes sense to
you now...